requirements.txt
streamlit
google-generativeai
python-dotenv
faiss-cpu
sentence-transformers
langchain>=0.2.0
langchain-core
langchain-community
langchain-huggingface
langchain-text-splitters
langchain_community
